requirements.txt
Mako==1.3.10
MarkupSafe==3.0.3
python-frontmatter==1.1.0
PyYAML==6.0.3
regex==2025.10.22
